Refactor: host config to src/ with plugin architecture
