+++ src/Host/Grpc/GreeterService.cs
@@ -1,0 +1,16 @@
+using Grpc.Core;
+
+namespace Host.Grpc;
+
+public class GreeterService(ILogger<GreeterService> logger) : Greeter.GreeterBase
+{
+    public override Task<HelloReply> SayHello(HelloRequest request, ServerCallContext context)
+    {
+        logger.LogInformation("The message is received from {Name}", request.Name);
+
+        return Task.FromResult(new HelloReply
+        {
+            Message = "Hello " + request.Name
+        });
+    }
+}
